Suppor for special float values
The `tuned_literal_eval` is an extension to the `ast` one that allows 
`float("nan")` and `float("inf")`. As it was taken of the dev version 
of the `ast` library, it also allows `set()`.

Because of this there is a vunerability if the `float`, `set`, `list`, 
etc, built-in functions are overwritten.